+++ packages/core/CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## [Unreleased]
+### Fixed
+- Added handling for `0xcf` in MsgPack as Columba uses it for NomadNet requests
 
 ## [0.4.3] - 2026-07-24
 ### Fixed